chore: remove failed runs from results (#14)
Adds a new step to the `process` job that cleans up benchmark
results for each combination run for all different package managers.

This way we can be sure that failed runs are not impacting result
numbers, while also allowing for a more resilient system that will
still be able to complete a full benchmark even if a few runs for
a specific package manager + fixture + variation fails to complete.

Fixes: https://github.com/vltpkg/benchmarks/issues/11